+++ bff/src/Altinn.Authentication.UI/Altinn.Authentication.UI.Core/SystemRegister/ISystemRegisterClient.cs
@@ -1,0 +1,6 @@
+﻿namespace Altinn.Authentication.UI.Core.SystemRegister;
+
+public interface ISystemRegisterClient
+{
+    Task<List<RegisteredSystemDTO>> GetListRegSys();
+}
