+++ bff/src/Altinn.Authentication.UI/Altinn.Authentication.UI.Core/SystemRegister/RegisteredSystemDTO.cs
@@ -1,0 +1,6 @@
+﻿namespace Altinn.Authentication.UI.Core.SystemRegister;
+
+public class RegisteredSystemDTO
+{
+
+}
