+++ bff/src/Altinn.Authentication.UI/Altinn.Authentication.UI.Core/SystemRegister/SystemRegisterService.cs
@@ -1,0 +1,20 @@
+﻿namespace Altinn.Authentication.UI.Core.SystemRegister;
+
+public class SystemRegisterService : ISystemRegisterService
+{
+    ISystemRegisterClient _systemRegisterClient;
+
+    public SystemRegisterService(ISystemRegisterClient systemRegisterClient)
+    {
+        _systemRegisterClient = systemRegisterClient;
+    }
+
+    public async Task<List<RegisteredSystemDTO>> GetListRegSys()
+    {
+        List<RegisteredSystemDTO> lista = new();
+
+        lista = await _systemRegisterClient.GetListRegSys();
+
+        return lista;
+    }
+}
