--- bff/src/Altinn.Authentication.UI/Altinn.Authentication.UI.Core/SystemRegisters/ISystemRegisterClient.cs
@@ -1,5 +1,0 @@
-﻿namespace Altinn.Authentication.UI.Core.SystemRegisters;
-
-public interface ISystemRegisterClient
-{
-}
