--- bff/src/Altinn.Authentication.UI/Altinn.Authentication.UI.Core/SystemRegisters/RegisteredSystemDTO.cs
@@ -1,6 +1,0 @@
-﻿namespace Altinn.Authentication.UI.Core.SystemRegisters;
-
-public class RegisteredSystemDTO
-{
-
-}
