--- bff/src/Altinn.Authentication.UI/Altinn.Authentication.UI.Core/SystemRegisters/SystemRegisterService.cs
@@ -1,11 +1,0 @@
-﻿namespace Altinn.Authentication.UI.Core.SystemRegisters;
-
-public class SystemRegisterService : ISystemRegisterService
-{
-    public async Task<List<RegisteredSystemDTO>> GetListRegSys()
-    {
-        List<RegisteredSystemDTO> lista = new();
-
-        return lista;
-    }
-}
